feat(assistant): Phase 7 hardening and widget refactor
Fix AI_InvalidPromptError caused by passing UIMessage[] directly to
streamText — now awaits convertToModelMessages() to produce the expected
ModelMessage[] format.

Harden the assistant panel: key status check on open (idle -> checking ->
set/unset), friendly error mapping for auth/rate-limit/provider failures,
streamText wrapped in try/catch with HTTP error codes (422/429/502), and
a system prompt directive for markdown recipe formatting.

Add react-markdown to render assistant messages with Tailwind-styled
components (bullets, numbered lists, bold) instead of raw text. Replace
the sparkle SVG on the floating button with the app logo.

Refactor the 400-line widget into three files: proposal-card.tsx
(confirmation UI), use-assistant.ts (state/mutations/handlers), and a
lean assistant-widget.tsx (rendering only).
